+++ app/channels/application_cable/connection.rb
@@ -1,0 +1,9 @@
+module ApplicationCable
+  class Connection < ActionCable::Connection::Base
+    identified_by :connection_token
+
+    def connect
+      self.connection_token = SecureRandom.uuid
+    end
+  end
+end