+++ app/helpers/application_helper.rb
@@ -1,5 +1,6 @@
 module ApplicationHelper
   include Pagy::Frontend
+  include Chartkick::Helper
 
   FLASH_CLASSES = {
     'notice' => 'success',